Make tag list return list of references
This is to accommodate "Lightweight tags" being keeping track of their
names. Now every value in the returned list will know the name that it
is being referenced by, whether it is referencing a commit or a real tag
object.

This is done so I can get the latest tag object for codeabode and still
know its name without passing around two objects.

I took a look at pygit2 for ideas on this as well, they suggest just
using repo.references to filter for the tags, which, as you can imagine,
returns reference objects